Reading workflows to populate export configs
This allows a workflow to be used in exports, allowing a study author
to keep all the relevant data in the workflow itself and not worry about
the manifest for managing exports.

As a byproduct of this, it breaks out the parsing logic into its own
module to prevent circular imports between the CountsBuilder and the
StudyManifest.
